services/rag/requirements.txt
# =============================================================================
# MindFu RAG Service Dependencies
# =============================================================================

# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# LangChain ecosystem
langchain==0.3.7
langchain-community==0.3.5
langchain-core==0.3.15
langchain-text-splitters==0.3.0

# Vector store
qdrant-client==1.11.3

# Embeddings
sentence-transformers==3.2.1

# Database
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.35
psycopg2-binary==2.9.9

# Redis
redis==5.1.1

# HTTP client
httpx==0.27.2
openai==1.52.0

# Document processing
pypdf==5.0.1
python-docx==1.1.2
beautifulsoup4==4.12.3
markdown==3.7

# Utilities
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1

# Monitoring
prometheus-client==0.21.0
